+++ tests/backend-storage-unit/src/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./test/test-templates";
+export * from "./shared";
 export * from "./StorageUnitTests";
