+++ tests/backend-storage-unit/src/shared/index.ts
@@ -2,7 +2,6 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export * from "./BindingsTests";
-export * from "./CommonTests";
+export * from "./test-templates/BindingsTests";
 export * from "./Constants";
 export * from "./Mocks";
