+++ tests/backend-storage-unit/src/test/ClientStorage.test.ts
@@ -7,7 +7,8 @@ import { Readable } from "stream";
 import { ClientStorage, TransferConfig } from "@itwin/object-storage-core";
 
 import { config } from "./Config";
-import { Constants, testRelativeDirectoryValidation } from "./test-templates";
+import { Constants } from "./Constants";
+import { testRelativeDirectoryValidation } from "./test-templates/CommonTests";
 
 const { clientStorage } = config;
