+++ tests/backend-storage-unit/src/test/Constants.ts
@@ -3,16 +3,6 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 const constants = {
-  validS3ServerStorageConfig: {
-    dependencyName: "",
-    baseUrl: "https://testBaseUrl.com",
-    region: "testRegion",
-    bucket: "testBucket",
-    accessKey: "testAccessKey",
-    secretKey: "testSecretKey",
-    roleArn: "testRoleArn",
-    stsBaseUrl: "https://testStsBaseUrl.com",
-  } as const,
   invalidObjectReference: {
     baseDirectory: "testBaseDirectory",
     relativeDirectory: "testDirectory1\\testDirectory2",
